[jsonify]: use jsonify helper for all responses

+++ pages/api/notes/[notesId].ts
@@ -2,6 +2,7 @@ import dbConnect from "../../../lib/server";
 import Express from "express";
 import Notes from "../../../models/Notes";
 import mongoose from "mongoose";
+import { jsonify } from "../../../helpers/backend/api/jsonify";
 
 export default async function handler(req: Express.Request, res: Express.Response) {
     const { method, query } = req;
@@ -11,21 +12,26 @@ export default async function handler(req: Express.Request, res: Express.Respons
     const notesId = new mongoose.Types.ObjectId(query.notesId as string)                
     const note = await Notes.findById(notesId)
 
+    if(!note){
+        res.status(400).json(jsonify({message: "No such note found"}, false))
+        return;
+    }
+
     switch (method) {
         case "GET":            
-            res.status(200).json({ status: "success", note: note })
+            res.status(200).json(jsonify(note))
             break;
         case "PUT":
             Object.assign(note, req.body);
             note.save();
-            res.status(200).json({ status: "success", note: note })
+            res.status(200).json(jsonify(note))
             break;
         case "DELETE":
             note.delete();
-            res.status(200).json({ status: "success" })
+            res.status(200).json(jsonify({ status: "success" }))
             break;
         default:
-            res.status(400).json({ status: "fail", message: "No such method found for this route "})
+            res.status(400).json(jsonify({message: "No such method found for this route"}, false))
             break;
     }
 }
